worker: move plugin tx generation to worker
